Fix: DB, Redis 연결 완료 후 서버 listen 상태되도록 수정
* DB, Redis 연결 시도 -> 서버 listen -> DB, Redis 연결완료가 아닌
* DB, Redis 연결 시도 -> 연결 완료 -> 서버 listen 순서가 되도록 수정

Resolves: #86
