+++ src/main.ts
@@ -1,7 +1,6 @@
 import "reflect-metadata";
-import { server } from "./server";
-import config from "./config";
+import { bootstrap } from "./bootstrap";
 
-server
-  .build()
-  .listen(config.port, () => console.log(`✅ server on ${config.port}`));
+(async () => {
+  await bootstrap();
+})();
